Move schema into separate file
